+++ apps/backoffice-v2/src/pages/Document/Document.page.tsx
@@ -1,12 +1,22 @@
-import { Case } from '../Entity/components/Case/Case';
 import { useDocumentLogic } from '@/pages/Document/hooks/useDocumentLogic/useDocumentLogic';
+import { Case } from '../Entity/components/Case/Case';
 
-export const Document = () => {
+interface IDocumentProps {
+  wrapperClassName?: string;
+}
+
+export const Document = ({ wrapperClassName }: IDocumentProps) => {
   const { documents, isLoading } = useDocumentLogic();
 
   if (isLoading) {
     return null;
   }
 
-  return <Case.Documents hideOpenExternalButton documents={documents} />;
+  return (
+    <Case.Documents
+      hideOpenExternalButton
+      documents={documents}
+      wrapperClassName={wrapperClassName}
+    />
+  );
 };